fix: #39 move type files out of routes
